+++ packages/metascraper-logo-favicon/src/index.js
@@ -97,29 +97,22 @@ const sizeSelectors = [
   { tag: 'meta[name*="msapplication" i]', attr: 'content' } // Windows 8, Internet Explorer 11 Tiles
 ]
 
-const firstReachable = async (domNodeSizes, gotOpts) => {
+const firstReachable = async (domNodeSizes, resolveFaviconUrl, gotOpts) => {
   for (const { url } of domNodeSizes) {
-    const response = await reachableUrl(url, gotOpts)
-    if (!reachableUrl.isReachable(response)) continue
-    const contentType = response.headers['content-type']
-
     const urlExtension = extension(url)
-
     const contentTypes = ALLOWED_EXTENSION_CONTENT_TYPES.find(
       ([ext]) => ext === urlExtension
     )
 
-    if (
-      contentTypes &&
-      (!isValidContenType(contentType, contentTypes[1]) ||
-        response.body.toString()[0] === '<')
-    ) { continue }
-
-    return response.url
+    const response = await resolveFaviconUrl(url, contentTypes, gotOpts)
+    if (response !== undefined) return response.url
   }
 }
 
-const pickBiggerSize = async (sizes, { gotOpts } = {}) => {
+const pickBiggerSize = async (
+  sizes,
+  { resolveFaviconUrl = defaultResolveFaviconUrl, gotOpts } = {}
+) => {
   const sorted = sizes.reduce(
     (acc, item) => {
       acc[item.size.square ? 'square' : 'nonSquare'].push(item)
@@ -129,29 +122,50 @@ const pickBiggerSize = async (sizes, { gotOpts } = {}) => {
   )
 
   return (
-    (await firstReachable(pickBiggerSize.sortBySize(sorted.square), gotOpts)) ||
-    (await firstReachable(pickBiggerSize.sortBySize(sorted.nonSquare), gotOpts))
+    (await firstReachable(
+      pickBiggerSize.sortBySize(sorted.square),
+      resolveFaviconUrl,
+      gotOpts
+    )) ||
+    (await firstReachable(
+      pickBiggerSize.sortBySize(sorted.nonSquare),
+      resolveFaviconUrl,
+      gotOpts
+    ))
   )
 }
 
 pickBiggerSize.sortBySize = collection =>
   orderBy(collection, ['size.priority'], ['desc'])
 
-const createFavicon = ([ext, contentTypes]) => {
+const defaultResolveFaviconUrl = async (faviconUrl, contentTypes, gotOpts) => {
+  const response = await reachableUrl(faviconUrl, gotOpts)
+  if (!reachableUrl.isReachable(response)) return undefined
+
+  const contentType = response.headers['content-type']
+
+  if (contentTypes && !isValidContenType(contentType, contentTypes)) {
+    return undefined
+  }
+
+  if (contentTypes && response.body.toString()[0] === '<') {
+    return undefined
+  }
+
+  return response
+}
+
+const createFavicon = (
+  [ext, contentTypes],
+  resolveFaviconUrl = defaultResolveFaviconUrl
+) => {
   return async (url, { gotOpts } = {}) => {
     const faviconUrl = logo(`/favicon.${ext}`, { url })
-    if (!faviconUrl) return undefined
-    const response = await reachableUrl(faviconUrl, gotOpts)
-    if (!reachableUrl.isReachable(response)) return undefined
-    const contentType = response.headers['content-type']
-
-    if (
-      contentTypes &&
-      (!isValidContenType(contentType, contentTypes) ||
-        response.body.toString()[0] === '<')
-    ) { return undefined }
-
-    return response.url
+    return faviconUrl
+      ? resolveFaviconUrl(faviconUrl, contentTypes, gotOpts).then(
+        response => response?.url
+      )
+      : undefined
   }
 }
 
@@ -163,10 +177,16 @@ const google = async (url, { gotOpts } = {}) => {
 google.url = (url, size = 128) =>
   `https://www.google.com/s2/favicons?domain_url=${url}&sz=${size}`
 
-const createGetLogo = ({ withGoogle, withFavicon, gotOpts, keyvOpts }) => {
+const createGetLogo = ({
+  gotOpts,
+  keyvOpts,
+  resolveFaviconUrl,
+  withFavicon,
+  withGoogle
+}) => {
   const getLogo = async url => {
     const providers = ALLOWED_EXTENSION_CONTENT_TYPES.map(
-      ext => withFavicon && createFavicon(ext)
+      ext => withFavicon && createFavicon(ext, resolveFaviconUrl)
     )
       .concat(withGoogle && google)
       .filter(Boolean)
@@ -201,20 +221,31 @@ const createRootFavicon = ({ getLogo, withRootFavicon = true } = {}) => {
 }
 
 module.exports = ({
-  google: withGoogle = true,
   favicon: withFavicon = true,
-  rootFavicon: withRootFavicon = true,
+  google: withGoogle = true,
   gotOpts,
   keyvOpts,
-  pickFn = pickBiggerSize
+  pickFn = pickBiggerSize,
+  resolveFaviconUrl = defaultResolveFaviconUrl,
+  rootFavicon: withRootFavicon = true
 } = {}) => {
-  const getLogo = createGetLogo({ withGoogle, withFavicon, gotOpts, keyvOpts })
+  const getLogo = createGetLogo({
+    gotOpts,
+    keyvOpts,
+    resolveFaviconUrl,
+    withFavicon,
+    withGoogle
+  })
   const rootFavicon = createRootFavicon({ getLogo, withRootFavicon })
   return {
     logo: [
       toLogo(async ($, url) => {
         const sizes = getSizes($, sizeSelectors, url)
-        return await pickFn(sizes, { gotOpts, pickBiggerSize })
+        return await pickFn(sizes, {
+          resolveFaviconUrl,
+          gotOpts,
+          pickBiggerSize
+        })
       }),
       ({ url }) => getLogo(normalizeUrl(url)),
       rootFavicon
@@ -227,3 +258,4 @@ module.exports.createFavicon = createFavicon
 module.exports.createRootFavicon = createRootFavicon
 module.exports.createGetLogo = createGetLogo
 module.exports.pickBiggerSize = pickBiggerSize
+module.exports.resolveFaviconUrl = defaultResolveFaviconUrl